split packages: Move ComplexFeatureConstants to org.geotools.data.complex.util
