+++ modules/extension/app-schema/app-schema/src/main/java/org/geotools/data/complex/config/NonFeatureTypeProxy.java
@@ -20,9 +20,9 @@ package org.geotools.data.complex.config;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
-import org.geotools.data.complex.ComplexFeatureConstants;
 import org.geotools.data.complex.FeatureTypeMapping;
 import org.geotools.data.complex.feature.type.ComplexFeatureTypeFactoryImpl;
+import org.geotools.data.complex.util.ComplexFeatureConstants;
 import org.opengis.feature.type.AttributeDescriptor;
 import org.opengis.feature.type.AttributeType;
 import org.opengis.feature.type.ComplexType;
