+++ modules/extension/app-schema/app-schema/src/main/java/org/geotools/data/complex/filter/XPath.java
@@ -27,11 +27,11 @@ import java.util.logging.Logger;
 import javax.xml.namespace.QName;
 import org.geotools.appschema.feature.AppSchemaAttributeBuilder;
 import org.geotools.data.complex.AbstractMappingFeatureIterator;
-import org.geotools.data.complex.ComplexFeatureConstants;
 import org.geotools.data.complex.config.NonFeatureTypeProxy;
 import org.geotools.data.complex.config.Types;
 import org.geotools.data.complex.feature.type.ComplexFeatureTypeFactoryImpl;
 import org.geotools.data.complex.feature.type.UniqueNameFeatureTypeFactoryImpl;
+import org.geotools.data.complex.util.ComplexFeatureConstants;
 import org.geotools.factory.CommonFactoryFinder;
 import org.geotools.feature.AttributeImpl;
 import org.geotools.feature.AttributeTypeBuilder;
