+++ modules/extension/app-schema/app-schema/src/test/java/org/geotools/appschema/util/ComplexAttributeConverterFactoryTest.java
@@ -21,7 +21,7 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Set;
 import junit.framework.TestCase;
-import org.geotools.data.complex.ComplexFeatureConstants;
+import org.geotools.data.complex.util.ComplexFeatureConstants;
 import org.geotools.data.util.ComplexAttributeConverterFactory;
 import org.geotools.feature.AttributeImpl;
 import org.geotools.feature.ComplexAttributeImpl;
