+++ modules/extension/app-schema/app-schema/src/test/java/org/geotools/data/complex/DefaultGeometryTest.java
@@ -15,6 +15,7 @@ import org.geotools.data.DataAccess;
 import org.geotools.data.DataAccessFinder;
 import org.geotools.data.FeatureSource;
 import org.geotools.data.complex.config.Types;
+import org.geotools.data.complex.util.ComplexFeatureConstants;
 import org.geotools.feature.FeatureCollection;
 import org.geotools.feature.FeatureIterator;
 import org.junit.BeforeClass;
