+++ modules/extension/complex/src/main/java/org/geotools/data/complex/util/ComplexFeatureConstants.java
@@ -15,7 +15,7 @@
  *    Lesser General Public License for more details.
  */
 
-package org.geotools.data.complex;
+package org.geotools.data.complex.util;
 
 import org.geotools.feature.NameImpl;
 import org.geotools.feature.type.AttributeDescriptorImpl;
