+++ modules/unsupported/wfs-ng/src/main/java/org/geotools/data/wfs/internal/parsers/XmlComplexFeatureParser.java
@@ -25,8 +25,8 @@ import java.util.HashMap;
 import java.util.Map;
 import javax.xml.namespace.QName;
 import org.geotools.data.DataSourceException;
-import org.geotools.data.complex.ComplexFeatureConstants;
 import org.geotools.data.complex.config.Types;
+import org.geotools.data.complex.util.ComplexFeatureConstants;
 import org.geotools.factory.CommonFactoryFinder;
 import org.geotools.feature.AttributeBuilder;
 import org.geotools.feature.AttributeImpl;
